MAYA-129682 - Move `isMaterialsScope()` into Utils.
Move `isMaterialsScope()` to a shared location, so that it can be reused by LookdevX.
