+++ lib/mayaUsd/ufe/Utils.h
@@ -104,6 +104,10 @@ inline std::string uniqueChildName(const PXR_NS::UsdPrim& parent, const std::str
 MAYAUSD_CORE_PUBLIC
 bool isAGatewayType(const std::string& mayaNodeType);
 
+//! Returns true if \p item is a materials scope.
+MAYAUSD_CORE_PUBLIC
+bool isMaterialsScope(const Ufe::SceneItem::Ptr& item);
+
 MAYAUSD_CORE_PUBLIC
 Ufe::Path dagPathToUfe(const MDagPath& dagPath);
